+++ tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/common/controller/Hello.kt
@@ -1,4 +1,4 @@
-package ru.thetenzou.tsoddservice.controller
+package ru.thetenzou.tsoddservice.common.controller
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
